fix: breakpoint line height, coordinate label clipping, and rename transcript service
- Extend breakpoint dashed line to full track height when genes have
  unequal transcript counts
- Auto-adjust coordinate label anchor near track edges to prevent clipping
- Rename ensemblTranscriptService to genomeNexusTranscriptService
